// models/ocoTableModel.tsx
export class OCOTableModel{
    "symbol": string;
    "price": number;
    "quantity": number;
    "percentage" : string;
    "lastPrice": string;
    "ocoPercentage": number;
    "ocoLimitPrice" : number;
    "ocoStopPrice" : number;
    "ocoLimitTotal" : number;
    "ocoStopTotal" : number;
    "ocoProfitLoss": number;
    "total" : number;
    
}